+++ src/components/Animation.js
